# pre-compute scaled frequency in write_touchstone